move items around so that they can be imported without tweetnacl
